Fix MinGW based build
Changes:
- fix src/jomlib/jobserver.cpp for 64bit build;
- rename files src/jomlib/processXXX -> src/jomlib/jomprocessXXX;
- fix CMake and QMake projects to build with MinGW.

Task-number: QTCREATORBUG-17709

+++ src/jomlib/jobserver.cpp
@@ -47,8 +47,8 @@ bool JobServer::start(int maxNumberOfJobs)
 {
     Q_ASSERT(m_environment);
 
-    const uint randomId = (FileTime::currentTime().internalRepresentation() % UINT_MAX)
-        ^ reinterpret_cast<uint>(&maxNumberOfJobs);
+    const quint64 randomId = (FileTime::currentTime().internalRepresentation() % UINT_MAX)
+        ^ reinterpret_cast<quint64>(&maxNumberOfJobs);
     const QString semaphoreKey = QLatin1String("jomsrv-")
             + QString::number(QCoreApplication::applicationPid()) + QLatin1Char('-')
             + QString::number(randomId);
